+++ src/SkyImages.jl
@@ -450,7 +450,7 @@ Returns:
 """
 function crop_blob_to_location(
   blob::Array{Image, 1},
-  width::Union{Float64, Int64},
+  width::Union{Float64, Int},
   wcs_center::Vector{Float64})
     @assert length(wcs_center) == 2
     @assert width > 0
@@ -567,7 +567,7 @@ Args:
 Returns:
   An array of tiles containing the image.
 """
-function break_image_into_tiles(img::Image, tile_width::Int64)
+function break_image_into_tiles(img::Image, tile_width::Int)
   WW = ceil(Int, img.W / tile_width)
   HH = ceil(Int, img.H / tile_width)
   ImageTile[ ImageTile(hh, ww, img, tile_width) for hh=1:HH, ww=1:WW ]
@@ -577,7 +577,7 @@ end
 """
 Break a blob into tiles.
 """
-function break_blob_into_tiles(blob::Blob, tile_width::Int64)
+function break_blob_into_tiles(blob::Blob, tile_width::Int)
   [ break_image_into_tiles(img, tile_width) for img in blob ]
 end
 
@@ -628,7 +628,7 @@ function local_source_candidates(
     Float64[patches[s].radius * maximum(abs(eig(patches[s].wcs_jacobian)[1]))
             for s=1:length(patches)];
 
-  candidates = fill(Int64[], size(tiles));
+  candidates = fill(Int[], size(tiles));
   for h=1:size(tiles)[1], w=1:size(tiles)[2]
     # Find the patches that are less than the radius plus diagonal from the
     # center of the tile.  These are candidates for having some
@@ -659,7 +659,7 @@ Returns:
 """
 function get_local_sources(tile::ImageTile, patches::Vector{SkyPatch})
 
-    tile_sources = Int64[]
+    tile_sources = Int[]
     tile_center = Float64[mean(tile.h_range), mean(tile.w_range)]
 
     for patch_index in 1:length(patches)
@@ -682,7 +682,7 @@ end
 """
 Return a vector of (h, w) indices of tiles that contain this source.
 """
-function find_source_tiles(s::Int64, b::Int64, mp::ModelParams)
+function find_source_tiles(s::Int, b::Int, mp::ModelParams)
   [ ind2sub(size(mp.tile_sources[b]), ind) for ind in
     find([ s in sources for sources in mp.tile_sources[b]]) ]
 end
@@ -704,7 +704,7 @@ Returns:
   from tiles that do not have this source will be marked as 0.0.
 """
 function stitch_object_tiles(
-    s::Int64, b::Int64, mp::ModelParams{Float64}, tiled_blob::TiledBlob;
+    s::Int, b::Int, mp::ModelParams{Float64}, tiled_blob::TiledBlob;
     predicted::Bool=false)
 
   H, W = size(tiled_blob[b])
